NuSMV; add properties as a seperate logic.
Provide a LTL and CTL logic API which NuSMV supports.

Circuit.t generation needs to include properties (as well
as outputs) so an appropriate circuit generator is added
which considers properties.  A new output called
`__the_atomic_propositions` is generated which includes
all the atomic propositions of the properties.

Based on this multiple CTLSPEC and/or LTLSPEC properties can
be generated for verification.